popup-list, admin-users ja popup-form tiedostot pienennetty ja viety tiedostoihin mitä nämä tiedostot käyttävät "päätiedostoina" main.js kerää sitten kaiken yhteen.
