+++ testisivu.html
@@ -19,10 +19,10 @@
         <p>&copy; 2025 Popups</p>
     </footer>
     
-    <script src="http://localhost:3000/popup-embed.js"></script>
+    <script src="http://localhost:3000/popup-embed.js" class="popup-embed dark:text-white"></script>
 <script>
   window.addEventListener('load', function() {
-    ShowPopup('67d289c6f2e874439dabbacd');
+    ShowPopup('67d555a0e21034f2edb075ec');
   });
 </script>
